Fixes after advice 29 march

+++ src/random.c
@@ -1,6 +1,6 @@
 #include "random.h"
 
-int32_t getRandomNumber()
+uint16_t getRandomNumber()
 {
     return rand() % 10000;
 }
